refactor: optimize DataTable for reusability and implement dynamic filtering for category page
- Refactored DataTable component to improve reusability across the application
- Added support for dynamic column filtering based on user input
- Enhanced state management for filters and sorting to allow more flexible table configurations
- Integrated the refactored DataTable into the category page with dynamic filtering options
