+++ src/app/dashboard/category/components/columns.jsx
@@ -1,40 +1,40 @@
 import Image from "next/image";
-import RowActions from "./rowActions";
+import CategoryRowActions from "./CategoryRowActions";
 
-export const columns = [
-    {
-        accessorKey: "image",
-        header: "Image",
-        cell: ({ row }) => {
-            const image = row.getValue("image");
+export const CategoryColumns = [
+  {
+    accessorKey: "image",
+    header: "Image",
+    cell: ({ row }) => {
+      const image = row.getValue("image");
 
-            return (
-                <div>
-                    <Image
-                        src={image}
-                        alt="Image"
-                        width={80}
-                        height={60}
-                        className="bg-gray-100 rounded-md"
-                    />
-                </div>
-            );
-        },
+      return (
+        <div>
+          <Image
+            src={image}
+            alt="Image"
+            width={80}
+            height={60}
+            className="bg-gray-100 rounded-md"
+          />
+        </div>
+      );
     },
-    {
-        accessorKey: "name",
-        header: "Category name",
-    },
-    {
-        accessorKey: "categoryDescription",
-        header: "Description",
-    },
-    {
-        id: "actions",
-        cell: ({ row }) => {
-            const category = row.original;
+  },
+  {
+    accessorKey: "name",
+    header: "Category name",
+  },
+  {
+    accessorKey: "categoryDescription",
+    header: "Description",
+  },
+  {
+    id: "actions",
+    cell: ({ row }) => {
+      const category = row.original;
 
-            return <RowActions category={category} />;
-        },
+      return <CategoryRowActions category={category} />;
     },
+  },
 ];
